python-sma/requirements.txt
sqlalchemy==2.0.40
psycopg2-binary==2.9.10
pytest==8.3.5
requests==2.32.3